+++ modules/img2img.py
@@ -208,7 +208,6 @@ def img2img(id_task: str, mode: int, prompt: str, negative_prompt: str, prompt_s
         p.scale_by = scale_by
     p.scripts = modules.scripts.scripts_img2img
     p.script_args = args
-    p.extra_generation_params['Resize mode'] = resize_mode
     if mask:
         p.extra_generation_params["Mask blur"] = mask_blur
         p.extra_generation_params["Mask alpha"] = mask_alpha